Rename 'aws/internal/service/cloudcontrolapi' to 'aws/internal/service/cloudcontrol'.
