+++ aws/data_source_aws_cloudcontrolapi_resource.go
@@ -9,7 +9,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrolapi/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrol/finder"
 )
 
 func dataSourceAwsCloudControlApiResource() *schema.Resource {
