+++ aws/internal/service/cloudcontrol/waiter/status.go
@@ -6,7 +6,7 @@ import (
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/cloudcontrolapi"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrolapi/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrol/finder"
 	"github.com/terraform-providers/terraform-provider-aws/aws/internal/tfresource"
 )
