+++ aws/resource_aws_cloudcontrolapi_resource.go
@@ -17,8 +17,8 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 	"github.com/mattbaird/jsonpatch"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrolapi/finder"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrolapi/waiter"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrol/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrol/waiter"
 	cffinder "github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudformation/finder"
 	"github.com/terraform-providers/terraform-provider-aws/aws/internal/tfresource"
 )
