+++ aws/resource_aws_cloudcontrolapi_resource_test.go
@@ -11,7 +11,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/acctest"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
-	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrolapi/finder"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/service/cloudcontrol/finder"
 	"github.com/terraform-providers/terraform-provider-aws/aws/internal/tfresource"
 )
